requirements.txt
Django==5.2.2
djangorestframework==3.16.0
Faker==37.3.0
